Classes for magnets, improve plots and overall sintax
The typing hinting structure of radiation_source.pyi was moved to radiation_source.py, where it is facilitated.

The scope of SynchrotronRadiation was better delimited. Now the sintax to define the magnetic fields were moved from it to the new file mag_elements.py, which includes undulators, bendings and field maps, with also useful methods for evaluate characteristic energies for each one.
In order to allow a better configuration and control of the radiation calculation, the calc wavefront now must be executed by the user, which also can modify the default precision parameters used in the source wavefront calculation.
The radiation_source now accepts just trajectories objects (SRWLPrtTrj) (with their parameters well defined) to calculate radiation.

Lens and Fresnel Zone Plate optical elements included in opt_elements.py.

Documentation for the main classes and functions written or updated.